Split DataSetsController and LabeledDataController
