+++ api/Controllers/DataSetsController.cs
@@ -2,9 +2,19 @@ using api.Model;
 using api.Services;
 using Mapster;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Routing;
 
 namespace api.Controllers;
 
+public class DataSetControllerAttribute : Attribute, IRouteTemplateProvider
+{
+    public string Template => "datasets/{dataSetId:int}/[controller]";
+    public int? Order => 2;
+    public string Name { get; set; }
+}
+
+[ApiController]
+[Route("[controller]")]
 public class DataSetsController : GenericCrudController<DataSet, DataSetDto>
 {
     private readonly ILabeledDataService _dataService;
@@ -29,34 +39,6 @@ public class DataSetsController : GenericCrudController<DataSet, DataSetDto>
         return Ok(modified);
     }
 
-    [HttpGet("{id:int}/data")]
-    public async Task<IActionResult> GetData(int id)
-    {
-        return Ok(await _dataService.Get(id));
-    }
-
-    [HttpPost("{id:int}/data")]
-    public async Task<IActionResult> PostData(int id, LabeledDataDto data)
-    {
-        var createdLabel = await _dataService.Create(data.Adapt<LabeledData>());
-        return Created($"/datasets/{id}/data/{createdLabel.Id}", createdLabel);
-    }
-
-    [HttpPut("{dataSetId:int}/data/{id:int}")]
-    public async Task<IActionResult> PutData(int dataSetId, int id, LabeledDataDto data)
-    {
-        data.DataSetId = dataSetId;
-        var labeledData = data.Adapt<LabeledData>();
-        labeledData.Id = id;
-        var modified = await _dataService.Update(labeledData);
-        return Ok(modified);
-    }
-
-    [HttpDelete("{dataSetId:int}/data/{id:int}")]
-    public async Task<IActionResult> DeleteData(int dataSetId, int id)
-    {
-        await _dataService.Delete(id);
-        return Ok();
-    }
+    
 
 }
